Graph tests - Update: test_subgraph(): Moved from test_backend.py to test_graph.py - Update test_graph.py: Activated tests - Delete: Deprecated get_connected_subgraph(). Has moved to concept_graph() - Update: Needed to refactor these tests to match the codebase changes. Still much broken, unwritten assertions; WIP.
