+++ test/test_backend/test_backend.py
@@ -339,134 +339,3 @@ class TestBackend(unittest.TestCase):
             {'concept_id': 4091006, 'standard_concept': 'S', 'vocabulary_id': 'SNOMED'},
             {'concept_id': 4052321, 'standard_concept': 'S', 'vocabulary_id': 'SNOMED'}].sort(key=key)
         self.assertEquals(csmi, expected)
-
-    # TODO: test is failing. fix
-    #  - The cause of the err is that subgraph() is a different function now. it used to call connected_subgraph_from_nodes(), which no longer exists.
-    #    Is this test still valid? Is there any func that takes a list of concept IDs and returns their parentage tuples?
-    @unittest.skip("May need to remove this test. I think the code it was testing got removed. Skipping failing/erroring test temporarily.")
-    def test_subgraph(self):
-        """Tests subgraph()"""
-        from backend.routes.graph import wholegraph  # a little slow, and only needed here
-        # Basic unit test for a simple connected graph without a complex hierarchy
-        edges1 = wholegraph([1738170, 1738171, 1738202, 1738203])
-        """
-        ┌────────────┬──────────────────────┬───────────┬───────────────┬────────────────────┬────┬──────────────┬─────┬────────────┬───────────────┬───────────┬─────────────────────┐
-        │ concept_id │ concept_name         │ domain_id │ vocabulary_id │  concept_class_id  │ sc │ concept_code │ inv │ domain_cnt │    domain     │ total_cnt │ distinct_person_cnt │
-        ├────────────┼──────────────────────┼───────────┼───────────────┼────────────────────┼────┼──────────────┼─────┼────────────┼───────────────┼───────────┼─────────────────────┤
-        │    1738170 │ lopinavir            │ Drug      │ RxNorm        │ Ingredient         │ S  │ 195088       │ ∅   │          1 │ drug_exposure │      2188 │ 142                 │
-        │    1738171 │ lopinavir 133 MG     │ Drug      │ RxNorm        │ Clinical Drug Comp │ S  │ 331536       │ ∅   │          0 │               │         0 │ 0                   │
-        │    1738202 │ lopinavir 80 MG / ML │ Drug      │ RxNorm        │ Clinical Drug Comp │ S  │ 331538       │ ∅   │          0 │               │         0 │ 0                   │
-        │    1738203 │ lopinavir 200 MG     │ Drug      │ RxNorm        │ Clinical Drug Comp │ S  │ 597727       │ ∅   │          0 │               │         0 │ 0                   │
-        └────────────┴──────────────────────┴───────────┴───────────────┴────────────────────┴────┴──────────────┴─────┴────────────┴───────────────┴───────────┴─────────────────────┘
-        prefer concepts that do have counts
-        """
-        self.assertEqual(edges1, [ ( "1738170", "1738171" ), ( "1738170", "1738202" ), ( "1738170", "1738203" ) ])
-
-        """ x
-        select * from concept_relationship_plus
-        where concept_id_1 in (1738170, 1738171, 1738202, 1738203)
-          and concept_id_2 in (1738170, 1738171, 1738202, 1738203)
-          and concept_id_1 != concept_id_2
-        order by 5;
-
-        ┌─────────────────┬──────────────┬────────────────────┬──────────────┬─────────────────┬─────────────────┬──────────────┬────────────────────┐
-        │ vocabulary_id_1 │ concept_id_1 │   concept_name_1   │ concept_code │ relationship_id │ vocabulary_id_2 │ concept_id_2 │   concept_name_2   │
-        ├─────────────────┼──────────────┼────────────────────┼──────────────┼─────────────────┼─────────────────┼──────────────┼────────────────────┤
-        │ RxNorm          │      1738171 │ lopinavir 133 MG   │ 331536       │ RxNorm has ing  │ RxNorm          │      1738170 │ lopinavir          │
-        │ RxNorm          │      1738202 │ lopinavir 80 MG/ML │ 331538       │ RxNorm has ing  │ RxNorm          │      1738170 │ lopinavir          │
-        │ RxNorm          │      1738203 │ lopinavir 200 MG   │ 597727       │ RxNorm has ing  │ RxNorm          │      1738170 │ lopinavir          │
-        │ RxNorm          │      1738170 │ lopinavir          │ 195088       │ RxNorm ing of   │ RxNorm          │      1738202 │ lopinavir 80 MG/ML │
-        │ RxNorm          │      1738170 │ lopinavir          │ 195088       │ RxNorm ing of   │ RxNorm          │      1738171 │ lopinavir 133 MG   │
-        │ RxNorm          │      1738170 │ lopinavir          │ 195088       │ RxNorm ing of   │ RxNorm          │      1738203 │ lopinavir 200 MG   │
-        └─────────────────┴──────────────┴────────────────────┴──────────────┴─────────────────┴─────────────────┴──────────────┴────────────────────┘
-        relationship between sources and targets:
-        ┌─────────────────┬────────────────────────┬─────────────────┬──────────────────┬─────────────────────────┬─────────────────────────┐
-        │ relationship_id │   relationship_name    │ is_hierarchical │ defines_ancestry │ reverse_relationship_id │ relationship_concept_id │
-        ├─────────────────┼────────────────────────┼─────────────────┼──────────────────┼─────────────────────────┼─────────────────────────┤
-        │ RxNorm ing of   │ Ingredient of (RxNorm) │               1 │                1 │ RxNorm has ing          │                44818817 │
-        └─────────────────┴────────────────────────┴─────────────────┴──────────────────┴─────────────────────────┴─────────────────────────┘
-        (code: select * from concept_relationship_plus where concept_id_1 = 1738170 and concept_id_2 = 1738171;
-               select * from relationship where relationship_id = 'RxNorm ing of';)
-        """
-
-        #Test for a concept set that fills in the gaps (i.e between child and grandparent)
-        edges2 = wholegraph([1738170, 19122186])
-        self.assertEqual(edges2, [ ( "1738170", "1738203" ), ( "1738203", "19122186" ) ])
-        """
-        ┌────────────┬────────────────────────────────────────────────┬───────────┬───────────────┬────────────────────┬──────────────────┬──────────────┬────────────────┬────────────┬───────────────┬───────────┬─────────────────────┐
-        │ concept_id │                  concept_name                  │ domain_id │ vocabulary_id │  concept_class_id  │ standard_concept │ concept_code │ invalid_reason │ domain_cnt │    domain     │ total_cnt │ distinct_person_cnt │
-        ├────────────┼────────────────────────────────────────────────┼───────────┼───────────────┼────────────────────┼──────────────────┼──────────────┼────────────────┼────────────┼───────────────┼───────────┼─────────────────────┤
-        │    1738170 │ lopinavir                                      │ Drug      │ RxNorm        │ Ingredient         │ S                │ 195088       │ ∅              │          1 │ drug_exposure │      2188 │ 142                 │
-        │    1738203 │ lopinavir 200 MG                               │ Drug      │ RxNorm        │ Clinical Drug Comp │ S                │ 597727       │ ∅              │          0 │               │         0 │ 0                   │
-        │   19122186 │ lopinavir 200 MG / ritonavir 50 MG Oral Tablet │ Drug      │ RxNorm        │ Clinical Drug      │ S                │ 597730       │ ∅              │          1 │ drug_exposure │      5789 │ 833                 │
-        └────────────┴────────────────────────────────────────────────┴───────────┴───────────────┴────────────────────┴──────────────────┴──────────────┴────────────────┴────────────┴───────────────┴───────────┴─────────────────────┘
-
-        No relationship appears for concept relationship table.
-        """
-
-        #Test for a more complex hierarchial relationship
-        edges3 = wholegraph([321588, 4027255, 316139, 43530856,
-                             45766164,
-                             4024552,
-                             # missing node 4027255,
-                             43530961])
-        """
-        ┌────────────┬──────────────────────────────────────────────────┬───────────┬───────────────┬──────────────────┬──────────────────┬──────────────┬────────────────┬────────────┬────────────────────────────────────┬───────────┬─────────────────────┐
-        │ concept_id │                   concept_name                   │ domain_id │ vocabulary_id │ concept_class_id │ standard_concept │ concept_code │ invalid_reason │ domain_cnt │               domain               │ total_cnt │ distinct_person_cnt │
-        ├────────────┼──────────────────────────────────────────────────┼───────────┼───────────────┼──────────────────┼──────────────────┼──────────────┼────────────────┼────────────┼────────────────────────────────────┼───────────┼─────────────────────┤
-        │     321588 │ Heart disease                                    │ Condition │ SNOMED        │ Clinical Finding │ S                │ 56265001     │ ∅              │          1 │ condition_occurrence               │   1067160 │ 290357              │
-        │     316139 │ Heart failure                                    │ Condition │ SNOMED        │ Clinical Finding │ S                │ 84114007     │ ∅              │          2 │ drug_exposure,condition_occurrence │   4105468 │ 20,613310           │
-        │   43530856 │ High risk of heart failure, stage B              │ Condition │ SNOMED        │ Clinical Finding │ S                │ 609389009    │ ∅              │          1 │ condition_occurrence               │        20 │ 20                  │
-        │   45766164 │ Heart failure with reduced ejection fraction     │ Condition │ SNOMED        │ Clinical Finding │ S                │ 703272007    │ ∅              │          1 │ condition_occurrence               │     10913 │ 3386                │
-        │    4024552 │ Disorder of cardiac function                     │ Condition │ SNOMED        │ Clinical Finding │ S                │ 105981003    │ ∅              │          1 │ condition_occurrence               │      1192 │ 421                 │
-        │    4027255 │ Structural disorder of heart                     │ Condition │ SNOMED        │ Clinical Finding │ S                │ 128599005    │ ∅              │          1 │ condition_occurrence               │       613 │ 231                 │
-        │   43530961 │ Induced termination of pregnancy complicated by …│ Condition │ SNOMED        │ Clinical Finding │ S                │ 609507007    │ ∅              │          1 │ condition_occurrence               │        20 │ 20                  │
-        │            │…cardiac failure                                  │           │               │                  │                  │              │                │            │                                    │           │                     │
-        └────────────┴──────────────────────────────────────────────────┴───────────┴───────────────┴──────────────────┴──────────────────┴──────────────┴────────────────┴────────────┴────────────────────────────────────┴───────────┴─────────────────────┘
-
-        "vocabulary_id_1"|"concept_id_1"|"concept_name_1"                                               |"concept_code"|"relationship_id"|"vocabulary_id_2"|"concept_id_2"|"concept_name_2"
-        SNOMED           |4024552       |Disorder of cardiac function                                   |105981003     |Is a             |SNOMED           |321588        |Heart disease
-        SNOMED           |4027255       |Structural disorder of heart                                   |128599005     |Is a             |SNOMED           |321588        |Heart disease
-        SNOMED           |43530961      |Induced termination of pregnancy complicated by cardiac failure|609507007     |Is a             |SNOMED           |316139        |Heart failure
-        SNOMED           |45766164      |Heart failure with reduced ejection fraction                   |703272007     |Is a             |SNOMED           |316139        |Heart failure
-        SNOMED           |316139        |Heart failure                                                  |84114007      |Is a             |SNOMED           |4024552       |Disorder of cardiac function
-        SNOMED           |43530856      |High risk of heart failure, stage B                            |609389009     |Is a             |SNOMED           |4027255       |Structural disorder of heart
-        SNOMED           |316139        |Heart failure                                                  |84114007      |Subsumes         |SNOMED           |45766164      |Heart failure with reduced ejection fraction
-        SNOMED           |321588        |Heart disease                                                  |56265001      |Subsumes         |SNOMED           |4024552       |Disorder of cardiac function
-        SNOMED           |4024552       |Disorder of cardiac function                                   |105981003     |Subsumes         |SNOMED           |316139        |Heart failure
-        SNOMED           |321588        |Heart disease                                                  |56265001      |Subsumes         |SNOMED           |4027255       |Structural disorder of heart
-        SNOMED           |4027255       |Structural disorder of heart                                   |128599005     |Subsumes         |SNOMED           |43530856      |High risk of heart failure, stage B
-        SNOMED           |316139        |Heart failure                                                  |84114007      |Subsumes         |SNOMED           |43530961      |Induced termination of pregnancy complicated by cardiac failure
-        """
-        self.assertEqual(edges3, [ ( "4024552", "316139" ), ( "316139", "43530961" ), ( "316139", "45766164" ),
-                             ( "321588", "4024552" ), ( "321588", "4027255" ), ( "4027255", "43530856" ) ] )
-
-        #Testing a relationship where a common ancestor is needed to connect the graph
-        edges3 = wholegraph([4027255, 43530856, 4024552, 316139, 45766164, 43530961])
-        """
-        ┌────────────┬──────────────────────────────────────────────────┬───────────┬───────────────┬──────────────────┬──────────────────┬──────────────┬────────────────┬────────────┬────────────────────────────────────┬───────────┬─────────────────────┐
-        │ concept_id │                   concept_name                   │ domain_id │ vocabulary_id │ concept_class_id │ standard_concept │ concept_code │ invalid_reason │ domain_cnt │               domain               │ total_cnt │ distinct_person_cnt │
-        ├────────────┼──────────────────────────────────────────────────┼───────────┼───────────────┼──────────────────┼──────────────────┼──────────────┼────────────────┼────────────┼────────────────────────────────────┼───────────┼─────────────────────┤
-        │     316139 │ Heart failure                                    │ Condition │ SNOMED        │ Clinical Finding │ S                │ 84114007     │ ∅              │          2 │ drug_exposure,condition_occurrence │   4105468 │ 20,613310           │
-        │     321588 │ Heart disease                                    │ Condition │ SNOMED        │ Clinical Finding │ S                │ 56265001     │ ∅              │          1 │ condition_occurrence               │   1067160 │ 290357              │
-        │    4024552 │ Disorder of cardiac function                     │ Condition │ SNOMED        │ Clinical Finding │ S                │ 105981003    │ ∅              │          1 │ condition_occurrence               │      1192 │ 421                 │
-        │    4027255 │ Structural disorder of heart                     │ Condition │ SNOMED        │ Clinical Finding │ S                │ 128599005    │ ∅              │          1 │ condition_occurrence               │       613 │ 231                 │
-        │   43530856 │ High risk of heart failure, stage B              │ Condition │ SNOMED        │ Clinical Finding │ S                │ 609389009    │ ∅              │          1 │ condition_occurrence               │        20 │ 20                  │
-        │   43530961 │ Induced termination of pregnancy complicated by …│ Condition │ SNOMED        │ Clinical Finding │ S                │ 609507007    │ ∅              │          1 │ condition_occurrence               │        20 │ 20                  │
-        │            │…cardiac failure                                  │           │               │                  │                  │              │                │            │                                    │           │                     │
-        │   45766164 │ Heart failure with reduced ejection fraction     │ Condition │ SNOMED        │ Clinical Finding │ S                │ 703272007    │ ∅              │          1 │ condition_occurrence               │     10913 │ 3386                │
-        └────────────┴──────────────────────────────────────────────────┴───────────┴───────────────┴──────────────────┴──────────────────┴──────────────┴────────────────┴────────────┴────────────────────────────────────┴───────────┴─────────────────────┘
-
-        "vocabulary_id_1"|"concept_id_1"|"concept_name_1"                                               |"concept_code"|"relationship_id"|"vocabulary_id_2"|"concept_id_2"|"concept_name_2"
-        SNOMED           |316139        |Heart failure                                                  |84114007      |Is a             |SNOMED           |4024552       |Disorder of cardiac function
-        SNOMED           |43530856      |High risk of heart failure, stage B                            |609389009     |Is a             |SNOMED           |4027255       |Structural disorder of heart
-        SNOMED           |45766164      |Heart failure with reduced ejection fraction                   |703272007     |Is a             |SNOMED           |316139        |Heart failure
-        SNOMED           |43530961      |Induced termination of pregnancy complicated by cardiac failure|609507007     |Is a             |SNOMED           |316139        |Heart failure
-        SNOMED           |4027255       |Structural disorder of heart                                   |128599005     |Subsumes         |SNOMED           |43530856      |High risk of heart failure, stage B
-        SNOMED           |316139        |Heart failure                                                  |84114007      |Subsumes         |SNOMED           |43530961      |Induced termination of pregnancy complicated by cardiac failure
-        SNOMED           |316139        |Heart failure                                                  |84114007      |Subsumes         |SNOMED           |45766164      |Heart failure with reduced ejection fraction
-        SNOMED           |4024552       |Disorder of cardiac function                                   |105981003     |Subsumes         |SNOMED           |316139        |Heart failure
-        """
-        self.assertEqual(edges3,
-                         [ ( "4024552", "316139" ), ( "316139", "43530961" ), ( "316139", "45766164" ),
-                           ( "321588", "4024552" ), ( "321588", "4027255" ), ( "4027255", "43530856" ) ] )
